Removed socket.io lines from server.js
Removed duplicate lines from server.js but still requiring as dependency. Note Chatrooms working using Handles, not as authenticated account. Would need help with this if we want Chatroom use/security based on authentication. Also not sure if we want to save chats to database or not ... up to y'all.